requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic settings
pydantic-settings==2.0.3

# Storage
aiofiles==23.2.1
boto3==1.34.0

# AI/ML
anthropic==0.25.0
openai==1.12.0
google-generativeai==0.4.0
httpx<0.25.0  # Pin httpx version for compatibility with AI libraries

# Document processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Optional dependencies for development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
